+++ test-clean/index.js
@@ -1,0 +1,15 @@
+// Test code
+// More code
+// Even more code
+// Change 1: Added new feature
+// Change 2: Fixed critical bug
+// Change 3: Documentation update
+// Change 4: Refactoring code
+// Change 5: Performance improvement
+// Change 6: Another bug fix
+// RCM ticket commit
+// Another RCM ticket commit
+// Third RCM ticket commit
+// Final RCM ticket commit
+// RCM-1005 commit
+// RCM-2999 commit